# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing donations and campaigns in the Dona Tutti platform",
        "title": "Dona Tutti API",
        "contact": {
            "name": "API Support",
            "email": "support@donatutti.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9999",
    "basePath": "/api",
    "paths": {
        "/api/donations/campaign/{id}": {
            "get": {
                "description": "Get a list of donations for a specific campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "List donations by campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/donation.Donation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            }
        },
...
        "/donations": {
            "get": {
                "description": "Get a list of all donations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "List all donations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/donation.Donation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new donation with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "Create a new donation",
                "parameters": [
                    {
                        "description": "Donation details",
                        "name": "donation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/donation.Donation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/donation.Donation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            }
        },
        "/donations/{id}": {
            "get": {
                "description": "Get donation details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "Get donation by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/donation.Donation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update donation details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "Update donation details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Donation details",
                        "name": "donation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/donation.Donation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            }
        },
        "/donors": {
            "get": {
                "description": "Get a list of all donors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donors"
                ],
                "summary": "List all donors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/donor.Donor"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new donor with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donors"
                ],
                "summary": "Create a new donor",
                "parameters": [
                    {
                        "description": "Donor details",
                        "name": "donor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/donor.Donor"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/donor.Donor"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            }
        },
        "/donors/{id}": {
            "get": {
                "description": "Get donor details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donors"
                ],
                "summary": "Get donor by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/donor.Donor"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update donor details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donors"
                ],
                "summary": "Update donor details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Donor details",
                        "name": "donor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/donor.Donor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "donation.Donation": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "campaign_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "donor_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_anonymous": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/donation.PaymentMethod"
                },
                "status": {
                    "$ref": "#/definitions/donation.DonationStatus"
                }
            }
        },
        "donation.DonationStatus": {
            "type": "string",
            "enum": [
                "completed",
                "pending",
                "failed",
                "refunded"
            ],
            "x-enum-varnames": [
                "DonationStatusCompleted",
                "DonationStatusPending",
                "DonationStatusFailed",
                "DonationStatusRefunded"
            ]
        },
        "donation.PaymentMethod": {
            "type": "string",
            "enum": [
                "MercadoPago",
                "Transferencia",
                "Efectivo",
                "Tarjeta"
            ],
            "x-enum-varnames": [
                "PaymentMethodMercadoPago",
                "PaymentMethodTransferencia",
                "PaymentMethodEfectivo",
                "PaymentMethodTarjeta"
            ]
        },
        "donor.Donor": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "errors.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request format"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}